added some test cases
The test case for the misbehavior of batch rules is broken,
as the tempering with the target makes An dirty, which leads
not to a rebuild of An, but to a rebuild of On with An being
dirty. But An was not reproduced by the batch rule and so
we use the messed with An ... which kind of sucks.